+++ ocean_model_skill_assessor/plot/line.py
@@ -2,12 +2,16 @@
 Time series plots.
 """
 
-# from matplotlib.pyplot import legend, subplots
-from typing import Union
+import pathlib
+from typing import Optional, Union
 
+import cf_pandas
+import cf_xarray
 import matplotlib.pyplot as plt
+import numpy as np
 
 from pandas import DataFrame
+from xarray import Dataset
 
 
 fs = 14
@@ -16,105 +20,65 @@ lw = 2
 col_model = "r"
 col_obs = "k"
 
-
 def plot(
-    df: DataFrame,
-    xname: Union[str, list],
-    yname: Union[str, list],
-    # reference: DataFrame,
-    # sample: DataFrame,
-    title: str,
-    xlabel: str = None,
-    ylabel: str = None,
-    figname: str = "figure.png",
+    obs: Union[DataFrame, Dataset],
+    model: Dataset,
+    xname: str,
+    yname: str,
+    title: Optional[str] = None,
+    xlabel: Optional[str] = None,
+    ylabel: Optional[str] = None,
+    figname: Union[str, pathlib.Path] = "figure.png",
     dpi: int = 100,
-    stats: dict = None,
     figsize: tuple = (15, 5),
+    return_plot: bool = False,
     **kwargs,
 ):
     """Plot time series or CTD profile.
 
-    Plot reference vs. sample as time series line plot.
+    Use for featuretype of timeSeries or profile.
+    Plot obs vs. model as time series line plot or CTD profile.
 
     Parameters
     ----------
-    reference: DataFrame
+    obs: DataFrame, Dataset
         Observation time series
-    sample: DataFrame
-        Model time series to compare against reference.
-    title: str
+    mode: Dataset
+        Model time series to compare against obs
+    xname : str
+        Name of variable to plot on x-axis when interpreted with cf-xarray and cf-pandas
+    yname : str
+        Name of variable to plot on y-axis when interpreted with cf-xarray and cf-pandas
+    title: str, optional
         Title for plot.
-    xlabel: str
+    xlabel: str, optional
         Label for x-axis.
-    ylabel: str
+    ylabel: str, optional
         Label for y-axis.
     figname: str
         Filename for figure (as absolute or relative path).
     dpi: int, optional
-        dpi for figure.
-    stats : dict, optional
-        Statistics describing comparison, output from `df.omsa.compute_stats`.
+        dpi for figure. Default is 100.
+    figsize : tuple, optional
+        Figsize to pass to `plt.figure()`. Default is (15,5).
+    return_plot : bool
+        If True, return plot. Use for testing.
     """
-    fig, ax = plt.subplots(1, 1, figsize=figsize)
-    # probably CTD profile plot (depth on y axis)
-    if isinstance(xname, list):
-        for name in xname:
-            if name == "obs":
-                label = "data"
-                color = col_obs
-            elif name == "model":
-                label = "model"
-                color = col_model
-            df.plot(
-                x=name,
-                y=yname,
-                ax=ax,
-                fontsize=fs,
-                lw=lw,
-                subplots=False,
-                label=label,
-                color=color,
-            )
-    # probably time series plot
-    elif isinstance(yname, list):
-        for name in yname:
-            if name == "obs":
-                label = "data"
-                color = col_obs
-            elif name == "model":
-                label = "model"
-                color = col_model
-            df.plot(
-                x=xname,
-                y=name,
-                ax=ax,
-                fontsize=fs,
-                lw=lw,
-                subplots=False,
-                label=label,
-                color=color,
-            )
-        ax.set_xlim(df[xname].min(), df[xname].max())
-    # df[xname].plot(ax=ax, label="observation", fontsize=fs, lw=lw, color=col_obs)
-    # df[yname].plot(ax=ax, label="model", fontsize=fs, lw=lw, color=col_model)
-    if stats is not None:
-        stat_sum = ""
-        types = ["bias", "corr", "ioa", "mse", "ss", "rmse"]
-        if "dist" in stats:
-            types += ["dist"]
-        for type in types:
-            stat_sum += f"{type}: {stats[type]['value']:.1f}  "
-            # add line mid title if tall plot instead of wide plot
-            if type == "ioa" and figsize[1] > figsize[0]:
-                stat_sum += "\n"
 
-        title = f"{title}: {stat_sum}"
+    fig, ax = plt.subplots(1, 1, figsize=figsize, layout="constrained")
+    ax.plot(obs.cf[xname], obs.cf[yname], label="data", lw=lw, color=col_obs)
+    ax.plot(np.array(model.cf[xname]), np.array(model.cf[yname]), label="model", lw=lw, color=col_model)
 
-    ax.set_title(title, fontsize=fs_title, loc="left")
+    plt.tick_params(axis="both", labelsize=fs)
+
+    ax.set_title(title, fontsize=fs_title, loc="left", wrap=True)
     if ylabel is not None:
         ax.set_ylabel(ylabel, fontsize=fs)
     if xlabel is not None:
         ax.set_xlabel(xlabel, fontsize=fs)
     plt.legend(loc="best")
 
-    fig.savefig(figname, dpi=dpi, bbox_inches="tight")
+    fig.savefig(figname, dpi=dpi,)#, bbox_inches="tight")
+    
+    if return_plot:
+        return fig