+++ tests/test_utils.py
@@ -1,5 +1,7 @@
+import pathlib
 from unittest import mock
 
+import cf_pandas
 import intake_xarray
 import numpy as np
 import pytest
@@ -8,6 +10,7 @@ import xarray as xr
 
 from intake.catalog import Catalog
 from intake.catalog.local import LocalCatalogEntry
+from unittest import TestCase
 
 import ocean_model_skill_assessor as omsa
 
@@ -41,10 +44,15 @@ ds["lon"] = (
 #     {"units": "degrees_east", "standard_name": "longitude"},
 # )
 
+@pytest.fixture(scope="session")
+def project_cache(tmp_path_factory):
+    directory = tmp_path_factory.mktemp("cache")
+    return directory
+
 
 @mock.patch("intake_xarray.base.DataSourceMixin.to_dask")
 @mock.patch("intake.open_catalog")
-def test_kwargs_search_from_model(mock_open_cat, mock_to_dask):
+def test_kwargs_search_from_model(mock_open_cat, mock_to_dask, project_cache):
 
     kwargs_search = {"model_name": "path", "project_name": "test_project"}
 
@@ -69,7 +77,8 @@ def test_kwargs_search_from_model(mock_open_cat, mock_to_dask):
 
     mock_to_dask.return_value = ds
 
-    kwargs_search = omsa.utils.kwargs_search_from_model(kwargs_search)
+    paths = omsa.paths.Paths(project_name="projectA", cache_dir=project_cache)
+    kwargs_search = omsa.utils.kwargs_search_from_model(kwargs_search, paths)
     output = {
         "min_lon": 0.0,
         "max_lon": 9.0,
@@ -86,7 +95,7 @@ def test_kwargs_search_from_model(mock_open_cat, mock_to_dask):
         "model_name": "path",
         "project_name": "test_project",
     }
-    kwargs_search = omsa.utils.kwargs_search_from_model(kwargs_search)
+    kwargs_search = omsa.utils.kwargs_search_from_model(kwargs_search, paths)
     output = {
         "min_lon": 0.0,
         "max_lon": 9.0,
@@ -105,7 +114,7 @@ def test_kwargs_search_from_model(mock_open_cat, mock_to_dask):
         "model_name": "path",
         "project_name": "test_project",
     }
-    kwargs_search = omsa.utils.kwargs_search_from_model(kwargs_search)
+    kwargs_search = omsa.utils.kwargs_search_from_model(kwargs_search, paths)
     output = {
         "min_lon": 1,
         "max_lon": 2,
@@ -127,11 +136,12 @@ def test_kwargs_search_from_model(mock_open_cat, mock_to_dask):
         "project_name": "test_project",
     }
     with pytest.raises(KeyError):
-        kwargs_search = omsa.utils.kwargs_search_from_model(kwargs_search)
+        kwargs_search = omsa.utils.kwargs_search_from_model(kwargs_search, paths)
 
 
 def test_find_bbox():
-    lonkey, latkey, bbox, p1 = omsa.utils.find_bbox(ds)
+    paths = omsa.paths.Paths(project_name="projectA", cache_dir=project_cache)
+    lonkey, latkey, bbox, p1 = omsa.utils.find_bbox(ds, paths)
 
     assert lonkey == "lon"
     assert latkey == "lat"
@@ -156,3 +166,27 @@ def test_shift_longitudes():
         {"units": "degrees_east", "standard_name": "longitude", "axis": "X"},
     )
     assert all(omsa.shift_longitudes(ds).cf["longitude"] == ds.cf["longitude"])
+
+
+@pytest.fixture(scope="session")
+def project_cache(tmp_path_factory):
+    directory = tmp_path_factory.mktemp("cache")
+    return directory
+
+
+def test_vocab(project_cache):
+    paths = omsa.paths.Paths(project_name="projectA", cache_dir=project_cache)
+    v1 = omsa.utils.open_vocabs("general", paths)
+    v2 = omsa.utils.open_vocabs(["general"], paths)
+    v3 = omsa.utils.open_vocabs(project_cache / pathlib.PurePath("vocab/general"), paths)
+    v4 = cf_pandas.Vocab(project_cache / pathlib.PurePath("vocab/general.json"))
+    TestCase().assertDictEqual(v1.vocab, v2.vocab)
+    TestCase().assertDictEqual(v1.vocab, v3.vocab)
+    TestCase().assertDictEqual(v1.vocab, v4.vocab)
+
+
+def test_vocab_labels(project_cache):
+    paths = omsa.paths.Paths(project_name="projectA", cache_dir=project_cache)
+    v1 = omsa.utils.open_vocab_labels("vocab_labels", paths)
+    v2 = omsa.utils.open_vocab_labels(project_cache / pathlib.PurePath("vocab/vocab_labels"), paths)
+    TestCase().assertDictEqual(v1, v2)